runtests: Add an option (-k|--keep-core) to keep a core file
Sometimes, when the Test Suite is being used to diagnose some particularly
heinous crash in Asterisk, it is useful to keep the core file generated from
the crashing instance. By default, the Test Suite removes these core files.
This is a good default option to have; the Test Suite is run on CI build
agents, and in the past we ran into situations where the generated core files
caused the disk to run out of memory (particularly when someone commits a
change that causes Asterisk to crash on startup/shutdown).

This patch adds a command line option, -k or --keep-core, that lets you
override this behaviour when running the Test Suite. This makes it a bit
easier to diagnose problems when running the Test Suite locally.

+++ runtests.py
@@ -196,7 +196,8 @@ class TestRun:
             finally:
                 dest_file.close()
                 try:
-                    os.unlink(core)
+                    if not self.options.keep_core:
+                        os.unlink(core)
                 except OSError, e:
                     print "Error removing core file: %s: " \
                           "Beware of the stale core file in CWD!" % (e,)
@@ -607,6 +608,9 @@ def main(argv=None):
     parser.add_option("-v", "--version",
                       dest="version", default=None,
                       help="Specify the version of Asterisk rather then detecting it.")
+    parser.add_option("-k", "--keep-core", action="store_true",
+                      dest="keep_core", default=False,
+                      help="Keep the 'core' file if Asterisk crashes.")
     parser.add_option("-L", "--list-tags", action="store_true",
                       dest="list_tags", default=False,
                       help="List available tags")
